Rename: 하단바 글 목록, 오타 수정(멘토링)
